Refactor 06-02-2024

+++ StartMenuEngine.cpp
@@ -13,8 +13,8 @@ StartMenuEngine::StartMenuEngine()
 	const Color idleColor(70, 70, 70);
 	const Color activeColor(20, 20, 100);
 	const Vector2f sizeButtons(300.f, 100.f);
-	const float middleWidth = (VideoMode::getDesktopMode().width / 2.0f);
-	const float middleHeight = (VideoMode::getDesktopMode().height / 2.0f);
+	const float middleWidth = (static_cast<float>(VideoMode::getDesktopMode().width) / 2.0f);
+	const float middleHeight = (static_cast<float>(VideoMode::getDesktopMode().height) / 2.0f);
 	const Vector2f startButtonLocation((middleWidth - (sizeButtons.x / 2.f)), middleHeight - 150.f);
 	const Vector2f exitButtonLocation((middleWidth - (sizeButtons.x / 2.f)), middleHeight + 50.f);
 	m_StartButton = Button(startButtonLocation, sizeButtons, Files::FONT_FILE,
